+++ CHANGELOG.md
@@ -29,6 +29,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Fixed
 - Print relative paths for dumps in WebDriverDiagnostics as a workaround for [JPERF-158].
 - Fix serialization of the `help` CLI argument.
+- Remove custom page load timeout. Decreases [JPERF-249] occurrence.
+
+[JPERF-249]: https://ecosystem.atlassian.net/browse/JPERF-249
 
 ### Deprecated
 - Deprecate the Kotlin-defaults-ridden `VirtualUserOptions` constructor.